Added scriptfor calculating volume of a box

+++ operations/exercise1.cpp
@@ -1,7 +1,7 @@
 #include <cmath>
 #include <iostream>
 
-int convert_celsius_to_fahrenheit(double celsius) {
+double convert_celsius_to_fahrenheit(double celsius) {
 
 	double const_var = (9.0/5);
